add article handler
