+++ src/scripts/comp.js
@@ -8,21 +8,6 @@ export const Comp = {
   },
   createDashboardContainer() {
     return `
-        <nav>This is the nav bar</nav>
-        <div id="formContainer">
-          <fieldset>
-          <label for="journalDate">Name</label>
-          <input type="text" name="eventName" id="event-name" required/>
-          <label for="journalConcepts">Date</label>
-          <input type="date" name="eventDate" id="event-date" required>
-          <label for="journalEntry">Location</label>
-          <input type="text" name="eventLocation" id="event-location" required>
-          <div class="eventcreate">
-            <input id="create-event" type="button" value="Create Event">
-          </div>
-        </fieldset>
-        </div>
-        <div id="listContainer"></div>
     <nav>
       <button id="events">Events</button>
       <button id="articles">Articles</button>
