+++ src/scripts/main.js
@@ -1,10 +1,11 @@
 import { API } from "./api.js";
 import { Comp } from "./comp.js";
 import { Action } from "./action.js";
+import { articleHandler } from "./articleHandler.js";
 
 let validationArray = [];
 
-API.getValues().then(array => {
+API.getValues("users").then(array => {
   validationArray = array;
 });
 
@@ -19,29 +20,6 @@ if (sessionStorage.length === 0) {
       Action.logIn();
     })
 } else {
-    Action.addToDom("#container", Comp.createDashboardContainer())
+  Action.addToDom("#container", Comp.createDashboardContainer())
+  articleHandler()
 }
-
-document.querySelector("#articles").addEventListener("click", () => {
-  document.querySelector("#formContainer").innerHTML = Comp.articleForm()
-  document.querySelector("#article-save").addEventListener("click", () => {
-    console.log("button click")
-    let articleTitle = document.querySelector("#articleFormTitle").value
-    let articleSynopsis = document.querySelector("#articleFormSynopsis").value
-    let articleUrl = document.querySelector("#articleFormUrl").value
-    let articleObj = ({
-      title: articleTitle,
-      synopsis: articleSynopsis,
-      url: articleUrl
-    })
-    API.postValue("articles",articleObj)
-    document.querySelector("#listContainer").innerHTML = ""
-    fetch("http://localhost:8088/articles")
-      .then(data => data.json())
-      .then(newData => {
-        newData.forEach(article => {
-          document.querySelector("#listContainer").innerHTML += Comp.article(article)
-        })
-      })
-  })
-})
